begining of support for greek admin language
This commit starts the greek language support for
the administration interface which also includes
some e-mail stuff. This commit adds the english files
to the repository which will be gradually translated.
It also moves the catalog language files so that now
the repo matches the directory structure of the opencart
installation.